Extract FFmpeg engine runtime helpers
engine.py still carried availability caches, validation helpers, output naming, progress parsing, thumbnail probing, position resolution, and stream helpers inline with the operation implementations. Moving those primitives into engine_runtime_utils.py creates a smaller compatibility facade while keeping the private helper names importable from mcp_video.engine.

Constraint: Existing tests and downstream callers import private helpers from mcp_video.engine, so engine.py must re-export them explicitly.

Rejected: Rewrite engine.py into operation modules in one pass | too broad and high-conflict for the first engine slice.

Confidence: high

Scope-risk: moderate

Directive: Preserve engine.py as the compatibility facade while splitting operation families in later PRs.

Tested: ruff check mcp_video/engine.py mcp_video/engine_runtime_utils.py tests/test_engine.py tests/test_server.py tests/test_client.py tests/test_e2e.py

Tested: private helper re-export smoke from mcp_video.engine

Tested: python3 -m pytest tests/test_engine.py tests/test_server.py tests/test_client.py tests/test_e2e.py -q --tb=short

Tested: python3 -m pytest tests/test_engine_advanced.py tests/test_red_team.py tests/test_adversarial_audit.py -q --tb=short
